backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.10.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# External Integrations
twilio==8.10.0
openai==1.3.0
httpx==0.25.1

# Sentiment Analysis
textblob==0.17.1

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.1
